+++ ublox-cellular/src/services/data/tcp_stack.rs
@@ -1,4 +1,4 @@
-use super::ssl::{SecurityProfileId, SSL};
+// use super::ssl::{SecurityProfileId, SSL};
 use super::DataService;
 use super::{
     socket::{Error as SocketError, SocketHandle, SocketSetItem, TcpSocket, TcpState},
@@ -55,9 +55,10 @@ where
             .map_err(Self::Error::from)?;
 
         if tcp.state() == TcpState::Created {
-            self.enable_ssl(*socket, SecurityProfileId(0))
-                .map_err(Self::Error::from)?;
-
+            /*
+                        self.enable_ssl(*socket, SecurityProfileId(0))
+                            .map_err(Self::Error::from)?;
+            */
             self.network
                 .send_internal(
                     &ConnectSocket {
